Various shell improvements (#2)
* Use a single command to start yggdrasild immediately
systemctl can both enable and start a service in a single command.

* Do not leak temporary files
This uses a trap to remove temporary files. This avoids leaking
temporary files in case of an early exit.

* Do not spawn subshells in errors
Backticks in shell spawn a subshell. In the first case it would actually
start an infinite loop since it would invoke itself.

* Do not use reserved exit code 2
Exit code 2 is reserved for misuse of shell builtins.

https://tldp.org/LDP/abs/html/exitcodes.html
* Remove SBINDIR variable

Any shell command knows how it was called since it's stored in $0. This
simplifies the logic and makes it portable. It also opens the path to
branding.

* Fail early
Bash supports an exit early mode using set -e. This means it stops
executing whenever a command fails.